==> R/02_data_analysis.qmd <==
---
title: "Data Analysis on Corpus"
---

# Kaggle Data

```{r}
library(tidyverse)
kaggle_spam <- read_csv(here("data/corpus/kaggle_spam.csv"))
```


TODO: read in Kaggle data
Small exploration  of Kaggle data:

Inspect data - it looks like these posts could be useful but they'll be insufficient alone - expected due to their age (early 2000s) and their domain (emails). Another issue may be their length and their processing. 
```{r, kaggle_spam_samples}
kaggle_spam %>%
  filter(label_text == "spam") %>%
  sample_n(10) %>%
  pull(text)
```


Spam tends to be slightly longer, have more hyperlinks (though the median is 0) and on average contain more numbers (mean 12.2 v 49.2), though the median is 7 vs 8 showing the skewing effect of the max numbers (one spam post had 15,885).
```{r, kaggle_spam_summaries}
kaggle_spam %>%
  mutate(number_counts = str_count(text, "NUMBER")) %>%
  summarise(
    min = min(number_counts),
    max = max(number_counts),
    mean = mean(number_counts),
    median = median(number_counts),
    .by = label
  )

kaggle_spam %>%
  mutate(number_counts = str_count(text, "hyperlink|url")) %>%
  summarise(
    min = min(number_counts),
    max = max(number_counts),
    mean = mean(number_counts),
    median = median(number_counts),
    .by = label
  )

kaggle_spam %>%
  mutate(string_length = str_length(text)) %>%
  summarise(
    min = min(string_length),
    max = max(string_length),
    mean = mean(string_length),
    median = median(string_length),
    .by = label
  )
```

# Edge Data
```{r}
# read in from 01_data_acquisition.qmd
edge_758_spam %>%
  filter(str_detect(message, "The following browsers are supported")) # 1.3k / 60k
```
